# specs/openapi.generated.yaml
openapi: 3.1.0
info:
  title: OverVault API
  version: 0.1.0
paths:
...
  /api/users:
    get:
      tags:
      - users
      summary: List Users
      operationId: list_users_api_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserOut'
                type: array
                title: Response List Users Api Users Get
      security:
      - HTTPBearer: []
  /api/users/{user_id}:
    get:
      tags:
      - users
      summary: Get User
      operationId: get_user_api_users__user_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/files/{file_id}/permissions:
    get:
      tags:
      - permissions
      summary: List Permissions
      operationId: list_permissions_api_files__file_id__permissions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Inactive
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionOut'
                title: Response List Permissions Api Files  File Id  Permissions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - permissions
      summary: Grant Permission
      operationId: grant_permission_api_files__file_id__permissions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/permissions/{permission_id}:
    delete:
      tags:
      - permissions
      summary: Revoke Permission
      operationId: revoke_permission_api_permissions__permission_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: string
          title: Permission Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GrantRequest:
      properties:
        grantee:
          type: string
          title: Grantee
        permission:
          $ref: '#/components/schemas/PermissionLevel'
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        signature:
          anyOf:
          - type: string
          - type: 'null'
          title: Signature
      type: object
      required:
      - grantee
      - permission
      title: GrantRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PermissionLevel:
      type: string
      enum:
      - read
      - write
      - manage
      title: PermissionLevel
    PermissionOut:
      properties:
        id:
          type: string
          title: Id
        file_id:
          type: string
          title: File Id
        grantee:
          type: string
          title: Grantee
        permission:
          $ref: '#/components/schemas/PermissionLevel'
        status:
          type: string
          title: Status
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        granted_by:
          type: string
          title: Granted By
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        revoked_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Revoked Reason
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - file_id
      - grantee
      - permission
      - status
      - granted_by
      - created_at
      title: PermissionOut
...
    Role:
      type: string
      enum:
      - employee
      - manager
      - admin
      - auditor
      title: Role
...
    UserOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        name:
          type: string
          title: Name
        role:
          $ref: '#/components/schemas/Role'
        wallet_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Address
      type: object
      required:
      - id
      - email
      - name
      - role
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer